+++ src/js/auth/containers/AuthApp.jsx
@@ -3,8 +3,8 @@ import { connect } from 'react-redux';
 
 import recordEvent from '../../../platform/monitoring/record-event';
 import LoadingIndicator from '../../common/components/LoadingIndicator';
-import { apiRequest } from '../../common/helpers/api';
-import environment from '../../common/helpers/environment';
+import { apiRequest } from '../../../platform/utilities/api';
+import environment from '../../../platform/utilities/environment';
 
 export class AuthApp extends React.Component {
   constructor(props) {
