+++ src/js/common/schemaform/save-in-progress/api.js
@@ -1,6 +1,6 @@
 import Raven from 'raven-js';
 import recordEvent from '../../../../platform/monitoring/record-event';
-import environment from '../../helpers/environment';
+import environment from '../../../../platform/utilities/environment';
 import { sanitizeForm } from '../../../../platform/forms/helpers';
 
 export function removeFormApi(formId) {
