+++ src/js/facility-locator/components/FacilityMap.jsx
@@ -1,5 +1,5 @@
 import { mapboxToken } from './MapboxClient';
-import environment from '../../common/helpers/environment';
+import environment from '../../../platform/utilities/environment';
 import React, { Component } from 'react';
 
 class HowToGetHere extends Component {
