+++ src/js/gi/config.js
@@ -1,4 +1,4 @@
-import environment from '../common/helpers/environment';
+import environment from '../../platform/utilities/environment';
 
 export const api = {
   url: `${environment.API_URL}/v0/gi`,
