+++ src/js/letters/utils/helpers.jsx
@@ -2,8 +2,8 @@
 import React from 'react';
 import Raven from 'raven-js';
 
-import { apiRequest as commonApiClient } from '../../common/helpers/api';
-import environment from '../../common/helpers/environment';
+import { apiRequest as commonApiClient } from '../../../platform/utilities/api';
+import environment from '../../../platform/utilities/environment';
 import { formatDateShort } from '../../../platform/utilities/date';
 import {
   AVAILABILITY_STATUSES,
