+++ src/js/personalization/dashboard-beta/actions/index.js
@@ -1,6 +1,6 @@
 import recordEvent from '../../../../platform/monitoring/record-event';
 import { removeFormApi } from '../../../common/schemaform/save-in-progress/api';
-import { apiRequest } from '../../../common/helpers/api';
+import { apiRequest } from '../../../../platform/utilities/api';
 import { getUserData } from '../../../common/helpers/login-helpers';
 
 export const UPDATE_PROFILE_FIELDS = 'UPDATE_PROFILE_FIELDS';
