+++ src/js/personalization/va-profile-beta/actions/index.js
@@ -1,4 +1,4 @@
-import { apiRequest } from '../../../common/helpers/api';
+import { apiRequest } from '../../../../platform/utilities/api';
 import { isValidEmail, isValidPhone } from '../../../../platform/forms/validations';
 
 export const OPEN_MODAL = 'OPEN_MODAL';
