+++ src/js/vic-v2/config/form.js
@@ -5,7 +5,7 @@ import fullSchemaVIC from 'vets-json-schema/dist/VIC-schema.json';
 import IntroductionPage from '../containers/IntroductionPage';
 import ConfirmationPage from '../containers/ConfirmationPage';
 import IdentityFieldsWarning from '../components/IdentityFieldsWarning';
-import asyncLoader from '../../../platform/utilities/react/asyncLoader';
+import asyncLoader from '../../../platform/utilities/ui/asyncLoader';
 import DD214Description from '../components/DD214Description';
 import PhotoDescription from '../components/PhotoDescription';
 import { prefillTransformer, submit, identityMatchesPrefill } from '../helpers';
