+++ src/platform/monitoring/sentry.js
@@ -4,7 +4,7 @@
 
 import Raven from 'raven-js';
 
-import environment from '../../js/common/helpers/environment';
+import environment from '../utilities/environment';
 
 // url check is necessary for e2e tests and local environments
 const trackErrors = environment.BASE_URL.indexOf('localhost') < 0;
