+++ src/platform/utilities/environment/index.js
